Screen Assets
I've included a selection of basic game screens.

I'm not sure what exact options you wanted in the option screen, so
I've held off on doing that one for now.

The example main menu file is to show you an idea I had for when
someone selects an option (the skate draws a line across the option).
I'm not sure if that's something that's easy to implement or not, but
I've put the files for that in there.

I've also added a team-selection screen player in a grey color scheme.
This way player doesn't get confused as to which team they should join
(if their player color had been drawn from the original two).

These were all just ideas for screens, critique them and let me know if
I should go a different direction or not. I feel these should be
suitable for V1 though (I have ideas on how to expand the team select
screen after more design direction/style decisions later [see:
http://gamesdbase.com/Media/SYSTEM/Arcade/Select/big/NFL_Blitz_-_1997_-_
Midway_Games.jpg].